Mashao_mask_GEN/requirements.txt
streamlit
opencv-python
torch
torchvision
numpy
Pillow
diffusers
transformers
torchmetrics
accelerate
matplotlib
